+++ src/types/types.ts
@@ -1051,6 +1051,42 @@ export interface NftContractNftsResponse {
 }
 
 /**
+ * OpenSea's metadata for an NFT collection.
+ */
+export interface OpenSeaCollectionMetadata {
+  /** The floor price of the NFT. */
+  floorPrice?: number;
+  /** The name of the collection on OpenSea. */
+  collectionName?: string;
+  /** The approval status of the collection on OpenSea. */
+  safelistRequestStatus?: OpenSeaSafelistRequestStatus;
+  /** The image URL determined by OpenSea. */
+  imageUrl?: string;
+  /** The description of the collection on OpenSea. */
+  description?: string;
+  /** The homepage of the collection as determined by OpenSea.*/
+  externalUrl?: string;
+  /** The Twitter handle of the collection. */
+  twitterUsername?: string;
+  /** The Discord URL of the collection. */
+  discordUrl?: string;
+  /** Timestamp of when the OpenSea metadata was last ingested by Alchemy. */
+  lastIngestedAt?: string;
+}
+
+/** An OpenSea collection's approval status. */
+export enum OpenSeaSafelistRequestStatus {
+  /** Verified collection. */
+  VERIFIED = 'verified',
+  /** Collections that are approved on open sea and can be found in search results. */
+  APPROVED = 'approved',
+  /** Collections that requested safelisting on OpenSea. */
+  REQUESTED = 'requested',
+  /** Brand new collections. */
+  NOT_REQUESTED = 'not_requested'
+}
+
+/**
  * The response object for the {@link findContractDeployer} function.
  *
  * @public