+++ test/unit/nft-api.test.ts
@@ -4,12 +4,12 @@ import MockAdapter from 'axios-mock-adapter';
 import {
   Alchemy,
   BaseNft,
-  fromHex,
   GetFloorPriceResponse,
   GetNftsForOwnerOptions,
   GetOwnersForContractWithTokenBalancesResponse,
   Nft,
   NftAttributesResponse,
+  NftContract,
   NftContractBaseNftsResponse,
   NftContractNftsResponse,
   NftExcludeFilters,
@@ -18,7 +18,8 @@ import {
   OwnedBaseNftsResponse,
   OwnedNft,
   OwnedNftsResponse,
-  RefreshState
+  RefreshState,
+  fromHex
 } from '../../src';
 import {
   RawGetBaseNftsForContractResponse,
@@ -26,7 +27,8 @@ import {
   RawGetNftsForContractResponse,
   RawGetNftsResponse,
   RawGetOwnersForContractWithTokenBalancesResponse,
-  RawNftAttributeRarity
+  RawNftAttributeRarity,
+  RawOpenSeaCollectionMetadata
 } from '../../src/internal/raw-interfaces';
 import {
   getNftContractFromRaw,
@@ -69,13 +71,25 @@ describe('NFT module', () => {
     const symbol = 'NCN';
     const totalSupply = '9999';
     const tokenType = NftTokenType.ERC721;
+    const openSea: RawOpenSeaCollectionMetadata = {
+      floorPrice: 2.2998,
+      collectionName: 'Collection Name',
+      safelistRequestStatus: 'verified',
+      imageUrl: 'http://image.url',
+      description: 'A sample description',
+      externalUrl: 'http://external.url',
+      twitterUsername: 'twitter-handle',
+      discordUrl: 'https://discord.gg/example',
+      lastIngestedAt: '2022-10-26T22:24:49.000Z'
+    };
 
     const rawNftContractResponse = createRawNftContract(
       address,
       tokenType,
       name,
       symbol,
-      totalSupply
+      totalSupply,
+      openSea
     );
     const expectedNftContract = getNftContractFromRaw(rawNftContractResponse);
 
@@ -83,21 +97,61 @@ describe('NFT module', () => {
       mock.onGet().reply(200, rawNftContractResponse);
     });
 
-    it('can be called with raw parameters', async () => {
-      await alchemy.nft.getContractMetadata(address);
+    function verifyNftContractMetadata(
+      actualNftContract: NftContract,
+      expectedNftContract: NftContract,
+      address: string,
+      name: string,
+      symbol: string,
+      totalSupply: string,
+      tokenType?: NftTokenType,
+      openSea?: RawOpenSeaCollectionMetadata
+    ) {
+      expect(actualNftContract).toEqual(expectedNftContract);
+
+      expect(actualNftContract.address).toEqual(address);
+      expect(actualNftContract.name).toEqual(name);
+      expect(actualNftContract.symbol).toEqual(symbol);
+      expect(actualNftContract.totalSupply).toEqual(totalSupply);
+      expect(actualNftContract.tokenType).toEqual(tokenType);
+      if (openSea) {
+        expect(actualNftContract.openSea?.floorPrice).toEqual(
+          openSea.floorPrice
+        );
+        expect(actualNftContract.openSea?.collectionName).toEqual(
+          openSea.collectionName
+        );
+        expect(actualNftContract.openSea?.safelistRequestStatus).toEqual(
+          openSea.safelistRequestStatus
+        );
+        expect(actualNftContract.openSea?.imageUrl).toEqual(openSea.imageUrl);
+        expect(actualNftContract.openSea?.description).toEqual(
+          openSea.description
+        );
+        expect(actualNftContract.openSea?.externalUrl).toEqual(
+          openSea.externalUrl
+        );
+        expect(actualNftContract.openSea?.twitterUsername).toEqual(
+          openSea.twitterUsername
+        );
+        expect(actualNftContract.openSea?.discordUrl).toEqual(
+          openSea.discordUrl
+        );
+        expect(actualNftContract.openSea?.lastIngestedAt).toEqual(
+          openSea.lastIngestedAt
+        );
+      }
 
       expect(mock.history.get.length).toEqual(1);
       expect(mock.history.get[0].params).toHaveProperty(
         'contractAddress',
         address
       );
-    });
+    }
 
     it('returns the api response in the expected format', async () => {
-      const response = await alchemy.nft.getContractMetadata(address);
-
       verifyNftContractMetadata(
-        response,
+        await alchemy.nft.getContractMetadata(address),
         expectedNftContract,
         address,
         name,