+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Added support for ENS resolution on several methods. You should now be able to pass in an ENS domain into any param that requires an owner address. 
 - Added the `CoreNamespace.resolveName()` and `CoreNamespace.lookupAddress()` methods to resolve and lookup ENS domains and their owner addresses.
 - Fixed an issue where the `Accept-Encoding` header was incorrectly included in requests that originated from the browser (#174).
+- Added the `openSea` response to `NftNamespace.getContractMetadata()` (#162). 
 
 ## 2.2.0